sysctl: Prefer ctl_table_header in proc_sysctl
This is a preparation commit to use ctl_table_header instead of
ctl_table as the element that is passed around in proc_sysctl.c.
This will become necessary when the size can no longer be calculated
by searching for an empty sentinel in the ctl_table array and
will live in the ctl_table_header.

Signed-off-by: Joel Granados <j.granados@samsung.com>