networking: rm sentinel element from ctl_table array
This is part of the greater effort to remove the empty elements
(sentinels) at the end of the ctl_table arrays.

Replace the lines where procname is set to NULL with the equivalent size
value. Removed the "-1" that adjusted for having an extra empty element
when looping over the ctl_table arrays.

Signed-off-by: Joel Granados <j.granados@samsung.com>