netfilter: rm sentinel element from ctl_table array
This is part of the greater effort to remove the empty elements
(sentinels) at the end of the ctl_table arrays.

We reduce the size of the ctl_table array by modifying the neigh_vars
struct size and the last element in the nf_ct_sysctl_index enumeration.

Care is taken to remove all zeroing of ctl_table elements; with the new
size arg they are not needed.

Signed-off-by: Joel Granados <j.granados@samsung.com>