aio: rm sentinel element from ctl_table array
This is part of the greater effort to remove the empty elements
(sentinels) from the end of the clt_table arrays.

Make sure that we create an empty directory for fs/verity when
CONFIG_FS_VERITY_BUILTIN_SIGNATURES is not defined. In this case we use
the register sysctl call that expects a size.

Signed-off-by: Joel Granados <j.granados@samsung.com>