network/llc: rm sentinel element from ctl_table array
Make sure that "net/llc/station" lives on as a permanently empty
directory. Remove llc_station_table as it is no longer needed to
register the path. Add an argument to register_sysctl_mount_point so
"net/llc/station" can be registered under the net sysctl tree instead of
the default one.

This is part of the greater effort to remove the empty elements
(sentinels) from the end of the ctl_table arrays.

Signed-off-by: Joel Granados <j.granados@samsung.com>