Initialize pasid array while doing an hwtp_allocation
iommu_report_device_fault expects a pasid array to be populated so with an
iommu_attach_handle. Add this handle when the idev is iopf_enabled. And
remove it when we release ownership of the group.

The attach_handle is in the dev->iommu_group->pasid_array. This means
that if two or more process setup bare metal iopf dma on the same
device, they will not be able to have two different handlers. They will
inevitably clash when using the Page Request Queue.

Signed-off-by: Joel Granados <j.granados@samsung.com>