Initialize pasid array while doing an iommufd_fault_domain_replace_dev
iommu_report_device_fault expects a pasid array to be populated with an
iommu_attach_handle. Add this handle after the iopf is enabled in the
idev; remove it when we release ownership of the group.

The attach_handle is in the dev->iommu_group->pasid_array. This means
that if two or more process setup bare metal iopf dma on the same
device, they will not be able to have two different handlers. They will
inevitably clash when using the Page Request Queue.

Signed-off-by: Joel Granados <j.granados@samsung.com>